chore: Pin node version in docker base images and bump nvmrc
We were using a floating version of node in our base images, which means
we didn't control when we updated the node version. Also, we were using
an outdated version for local development, which didn't match the one
used for the base images.

This changeset bumps nvmrc to latest node (18.19.0) and pins down the
version of all node base images across dockerfiles in the project.